+++ Sources/SphericalGeometry/Utils.swift
@@ -8,11 +8,11 @@
 import Foundation
 
 class Utils {
-  class func toDegrees(radians: Float) -> Float {
-    return (radians * 180) / Float.pi
+  class func toDegrees(radians: Double) -> Double {
+    return (radians * 180) / Double.pi
   }
 
-  class func toRadians(angleDegrees: Float) -> Float {
-    return (angleDegrees * Float.pi) / 180.0
+  class func toRadians(angleDegrees: Double) -> Double {
+    return (angleDegrees * Double.pi) / 180.0
   }
 }
